// SDKs/twilio.com/api/1.36.0/typescript/src/sdk/models/operations/listsipipaccesscontrollist.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import * as shared from "../shared";
export const LISTSIPIPACCESSCONTROLLIST_SERVERS = [
	"https://api.twilio.com",
];



export class ListSipIpAccessControlListPathParams extends SpeakeasyBase {
  @Metadata({ data: "pathParam, style=simple;explode=false;name=AccountSid" })
  accountSid: string;
}


export class ListSipIpAccessControlListQueryParams extends SpeakeasyBase {
  @Metadata({ data: "queryParam, style=form;explode=true;name=PageSize" })
  pageSize?: number;
}


export class ListSipIpAccessControlListSecurity extends SpeakeasyBase {
  @Metadata({ data: "security, scheme=true;type=http;subtype=basic" })
  accountSidAuthToken: shared.SchemeAccountSidAuthToken;
}


export class ListSipIpAccessControlListRequest extends SpeakeasyBase {
  @Metadata()
  serverUrl?: string;

  @Metadata()
  pathParams: ListSipIpAccessControlListPathParams;

  @Metadata()
  queryParams: ListSipIpAccessControlListQueryParams;

  @Metadata()
  security: ListSipIpAccessControlListSecurity;
}


export class ListSipIpAccessControlListListSipIpAccessControlListResponse extends SpeakeasyBase {
  @Metadata({ data: "json, name=end" })
  end?: number;

  @Metadata({ data: "json, name=first_page_uri" })
  firstPageUri?: string;

  @Metadata({ data: "json, name=ip_access_control_lists", elemType: shared.ApiV2010AccountSipSipIpAccessControlList })
  ipAccessControlLists?: shared.ApiV2010AccountSipSipIpAccessControlList[];

  @Metadata({ data: "json, name=next_page_uri" })
  nextPageUri?: string;

  @Metadata({ data: "json, name=page" })
  page?: number;

  @Metadata({ data: "json, name=page_size" })
  pageSize?: number;

  @Metadata({ data: "json, name=previous_page_uri" })
  previousPageUri?: string;

  @Metadata({ data: "json, name=start" })
  start?: number;

  @Metadata({ data: "json, name=uri" })
  uri?: string;
}


export class ListSipIpAccessControlListResponse extends SpeakeasyBase {
  @Metadata()
  contentType: string;

  @Metadata()
  listSipIpAccessControlListResponse?: ListSipIpAccessControlListListSipIpAccessControlListResponse;

  @Metadata()
  statusCode: number;
}
